Be a little smarter about how we pass parameters.
A different approach: wherever possible, just pass the
entire params hash to the action.  That way, we don't
have to modify this code every time an Action class
gets a new attribute.

The implication is that the parameters to the web interface
will have the same names and semantics as the attributes
on the Pinto API.  But that is probably a good thing
anyway.