Finally got something that works.
Rewrote the tester to use Plack::Test, instead of my
own homegrown tester.

When running Pinto, we need to trap exceptions that
happen before the Action even starts (like using a
bogus Action name.  Unfortunatley, this means
the ActionResponder also needs to be able to write
to the output handle, to report such exceptions.

I'm not really happy with all this.  I just want
all exceptions to go straight back to client.