Update to use latest Pinto.
Pinto now uses static files for the index of each stack.
This eliminates the need for the server to generate a dynamic
index file upon request.  It also simplifes the server overall,
since it now only handles two kinds of requests: GET requests
for files, and POST requests for actions.